Rename option_parser.rb to parser.rb to match class name.
