+++ lib/rspec/core.rb
@@ -14,7 +14,7 @@ require 'rspec/core/pending'
 require 'rspec/core/world'
 require 'rspec/core/configuration'
 require 'rspec/core/command_line_configuration'
-require 'rspec/core/option_parser'
+require 'rspec/core/parser'
 require 'rspec/core/configuration_options'
 require 'rspec/core/command_line'
 require 'rspec/core/drb_command_line'
